Agentic AI/aether-search/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain_groq
duckduckgo-search
numpy
requests
trafilatura
sentence-transformers
